refactor: tests broken out to individual test files
